LPS-85528 Separate model classes in different files
